Set up new project and add essential files for portfolio site
Established the necessary groundwork for a new portfolio site. This includes the addition of key files such as ESLint, gitignore, App typescript and CSS files, Hero, Experience, CompetencePro components, among others. Additionally, the package.json and package-lock.json files were updated with the installation of required dependencies for the project.